// src/api/generated/models/TrendingChannelItem.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
/**
 * Trending channel item
 */
export type TrendingChannelItem = {
    /**
     * Channel ID
     */
    id: string;
    /**
     * Channel name
     */
    name: string;
    /**
     * Channel description
     */
    description?: (string | null);
    /**
     * Channel thumbnail
     */
    thumbnail_url?: (string | null);
    /**
     * Subscribers count
     */
    subscribers_count?: number;
    /**
     * Contents count
     */
    contents_count?: number;
};